+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# unreleased
+## Fixes
+* Re-lookup import deck cards when VERSION is out of date
+* Save deck automatically after import
+
 # 0.7.0
 ## Features
 * Allow grouping by no group